+++ edgeless_api/src/grpc_impl/inner/node_registration.rs
@@ -312,8 +312,32 @@ fn parse_function_log_entry(api_instance: &crate::grpc_impl::api::FunctionLogEnt
 
 fn parse_node_performance_samples(api_instance: &crate::grpc_impl::api::NodePerformanceSamples) -> crate::node_registration::NodePerformanceSamples {
     crate::node_registration::NodePerformanceSamples {
+        function_instantiate_times: api_instance
+            .function_instantiate_times
+            .iter()
+            .filter_map(|x| match uuid::Uuid::from_str(&x.id) {
+                Ok(val) => Some((val, x.samples.iter().map(parse_sample).collect())),
+                _ => None,
+            })
+            .collect(),
+        function_init_times: api_instance
+            .function_init_times
+            .iter()
+            .filter_map(|x| match uuid::Uuid::from_str(&x.id) {
+                Ok(val) => Some((val, x.samples.iter().map(parse_sample).collect())),
+                _ => None,
+            })
+            .collect(),
         function_execution_times: api_instance
             .function_execution_times
+            .iter()
+            .filter_map(|x| match uuid::Uuid::from_str(&x.id) {
+                Ok(val) => Some((val, x.samples.iter().map(parse_sample).collect())),
+                _ => None,
+            })
+            .collect(),
+        function_stop_times: api_instance
+            .function_stop_times
             .iter()
             .filter_map(|x| match uuid::Uuid::from_str(&x.id) {
                 Ok(val) => Some((val, x.samples.iter().map(parse_sample).collect())),
@@ -358,8 +382,32 @@ fn serialize_function_log_entry(req: &crate::node_registration::FunctionLogEntry
 
 fn serialize_node_performance_samples(req: &crate::node_registration::NodePerformanceSamples) -> crate::grpc_impl::api::NodePerformanceSamples {
     crate::grpc_impl::api::NodePerformanceSamples {
+        function_instantiate_times: req
+            .function_instantiate_times
+            .iter()
+            .map(|(id, samples)| crate::grpc_impl::api::Samples {
+                id: id.to_string(),
+                samples: samples.iter().map(serialize_sample).collect(),
+            })
+            .collect(),
+        function_init_times: req
+            .function_init_times
+            .iter()
+            .map(|(id, samples)| crate::grpc_impl::api::Samples {
+                id: id.to_string(),
+                samples: samples.iter().map(serialize_sample).collect(),
+            })
+            .collect(),
         function_execution_times: req
             .function_execution_times
+            .iter()
+            .map(|(id, samples)| crate::grpc_impl::api::Samples {
+                id: id.to_string(),
+                samples: samples.iter().map(serialize_sample).collect(),
+            })
+            .collect(),
+        function_stop_times: req
+            .function_stop_times
             .iter()
             .map(|(id, samples)| crate::grpc_impl::api::Samples {
                 id: id.to_string(),
@@ -470,13 +518,28 @@ mod test {
                 active_power: 25,
             },
             performance_samples: NodePerformanceSamples {
+                function_instantiate_times: std::collections::HashMap::from([
+                    (uuid::Uuid::new_v4(), vec![new_sample(101.0), new_sample(2.5), new_sample(3.0)]),
+                    (uuid::Uuid::new_v4(), vec![]),
+                    (uuid::Uuid::new_v4(), vec![new_sample(0.1), new_sample(0.2), new_sample(999.0)]),
+                ]),
+                function_init_times: std::collections::HashMap::from([
+                    (uuid::Uuid::new_v4(), vec![new_sample(201.0), new_sample(2.5), new_sample(3.0)]),
+                    (uuid::Uuid::new_v4(), vec![]),
+                    (uuid::Uuid::new_v4(), vec![new_sample(0.1), new_sample(0.2), new_sample(999.0)]),
+                ]),
                 function_execution_times: std::collections::HashMap::from([
-                    (uuid::Uuid::new_v4(), vec![new_sample(1.0), new_sample(2.5), new_sample(3.0)]),
+                    (uuid::Uuid::new_v4(), vec![new_sample(301.0), new_sample(2.5), new_sample(3.0)]),
+                    (uuid::Uuid::new_v4(), vec![]),
+                    (uuid::Uuid::new_v4(), vec![new_sample(0.1), new_sample(0.2), new_sample(999.0)]),
+                ]),
+                function_stop_times: std::collections::HashMap::from([
+                    (uuid::Uuid::new_v4(), vec![new_sample(401.0), new_sample(2.5), new_sample(3.0)]),
                     (uuid::Uuid::new_v4(), vec![]),
                     (uuid::Uuid::new_v4(), vec![new_sample(0.1), new_sample(0.2), new_sample(999.0)]),
                 ]),
                 function_transfer_times: std::collections::HashMap::from([
-                    (uuid::Uuid::new_v4(), vec![new_sample(1.0), new_sample(2.5), new_sample(3.0)]),
+                    (uuid::Uuid::new_v4(), vec![new_sample(501.0), new_sample(2.5), new_sample(3.0)]),
                     (uuid::Uuid::new_v4(), vec![]),
                     (uuid::Uuid::new_v4(), vec![new_sample(0.1), new_sample(0.2), new_sample(999.0)]),
                 ]),